+++ java-service-processor/src/main/java/internal/nbbrd/service/ModuleInfoEntries.java
@@ -20,18 +20,21 @@ import com.github.javaparser.ParseProblemException;
 import com.github.javaparser.StaticJavaParser;
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.NodeList;
-import com.github.javaparser.ast.modules.ModuleDeclaration;
 import com.github.javaparser.ast.modules.ModuleDirective;
-import com.github.javaparser.ast.nodeTypes.NodeWithName;
 
 import javax.annotation.processing.Filer;
 import javax.annotation.processing.FilerException;
+import javax.lang.model.element.Name;
+import javax.lang.model.element.PackageElement;
+import javax.lang.model.element.TypeElement;
+import javax.lang.model.util.Elements;
 import javax.tools.FileObject;
 import javax.tools.StandardLocation;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.NoSuchFileException;
 import java.util.*;
+import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
 
@@ -60,32 +63,62 @@ public class ModuleInfoEntries {
         }
     }
 
-    public static Optional<ModuleInfoEntries> parse(Filer filer) throws IOException {
+    public static Optional<ModuleInfoEntries> parse(Filer filer, Elements elements) throws IOException {
         try {
             FileObject src = filer.getResource(StandardLocation.SOURCE_PATH, "", "module-info.java");
-            return Optional.of(parse(src.getCharContent(false)));
+            return parse(src.getCharContent(false), PackageExpander.of(elements));
         } catch (FileNotFoundException | NoSuchFileException | FilerException | RuntimeException ex) {
             // ignore
             return Optional.empty();
         }
     }
 
-    static ModuleInfoEntries parse(CharSequence content) {
-        try {
-            CompilationUnit compilationUnit = StaticJavaParser.parse(content.toString());
-            List<String> imports = compilationUnit.getImports().stream().map(NodeWithName::getNameAsString).collect(toList());
+    interface PackageExpander {
 
-            Optional<ModuleDeclaration> module = compilationUnit.getModule();
-            if (module.isPresent()) {
-                return parse(imports, module.get().getDirectives());
-            }
-        } catch (ParseProblemException ex) {
-            // ignore
+        Stream<String> getEnclosedTypeNames(String packageName);
+
+        static PackageExpander of(Elements elements) {
+            return packageName -> {
+                PackageElement packageElement = elements.getPackageElement(packageName);
+                return packageElement != null
+                        ? packageElement
+                        .getEnclosedElements()
+                        .stream()
+                        .filter(TypeElement.class::isInstance)
+                        .map(TypeElement.class::cast)
+                        .map(TypeElement::getQualifiedName)
+                        .map(Name::toString)
+                        : Stream.empty();
+            };
         }
-        return ModuleInfoEntries.builder().build();
     }
 
-    private static ModuleInfoEntries parse(List<String> imports, NodeList<ModuleDirective> directives) {
+    static Optional<ModuleInfoEntries> parse(CharSequence moduleInfoContent, PackageExpander expander) {
+        try {
+            return parse(StaticJavaParser.parse(moduleInfoContent.toString()), expander);
+        } catch (ParseProblemException ex) {
+            // ignore
+            return Optional.empty();
+        }
+    }
+
+    private static Optional<ModuleInfoEntries> parse(CompilationUnit compilationUnit, PackageExpander expander) {
+        List<String> imports = parseImports(compilationUnit, expander);
+        return compilationUnit.getModule()
+                .map(moduleDeclaration -> parseDirectives(imports, moduleDeclaration.getDirectives()));
+    }
+
+    private static List<String> parseImports(CompilationUnit compilationUnit, PackageExpander expander) {
+        return compilationUnit
+                .getImports()
+                .stream()
+                .flatMap(importDeclaration -> importDeclaration.isAsterisk()
+                        ? expander.getEnclosedTypeNames(importDeclaration.getNameAsString())
+                        : Stream.of(importDeclaration.getNameAsString()))
+                .collect(toList());
+    }
+
+    private static ModuleInfoEntries parseDirectives(List<String> imports, NodeList<ModuleDirective> directives) {
         ModuleInfoEntries.Builder result = ModuleInfoEntries.builder();
         directives.forEach(directive -> {
             directive.ifModuleProvidesDirective(provides -> provides.getWith().forEach(z -> result.provision(resolveTypeName(imports, provides.getNameAsString()), resolveTypeName(imports, z.asString()))));